Use fetchData as props instead
